fix circular dep in tests
