+++ core/types/transaction_test.go
@@ -32,20 +32,16 @@ import (
 	"testing"
 	"time"
 
-	gokzg4844 "github.com/crate-crypto/go-kzg-4844"
 	"github.com/holiman/uint256"
 	"github.com/stretchr/testify/assert"
 
 	libcommon "github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/common/fixedgas"
-	"github.com/erigontech/erigon-lib/common/hexutility"
 	"github.com/erigontech/erigon-lib/crypto"
-	"github.com/erigontech/erigon-lib/crypto/kzg"
-	libtypes "github.com/erigontech/erigon-lib/types"
 	"github.com/erigontech/erigon/common"
 	"github.com/erigontech/erigon/common/u256"
+	"github.com/erigontech/erigon/core/types/typestest"
 	"github.com/erigontech/erigon/rlp"
-	"github.com/erigontech/erigon/txnprovider/txpool"
 )
 
 // The values in those tests are from the Transaction Tests
@@ -378,7 +374,7 @@ func TestTransactionCoding(t *testing.T) {
 		signer    = LatestSignerForChainID(libcommon.Big1)
 		addr      = libcommon.HexToAddress("0x0000000000000000000000000000000000000001")
 		recipient = libcommon.HexToAddress("095e7baea6a6c7c4c2dfeb977efac326af552d87")
-		accesses  = libtypes.AccessList{{Address: addr, StorageKeys: []libcommon.Hash{{0}}}}
+		accesses  = AccessList{{Address: addr, StorageKeys: []libcommon.Hash{{0}}}}
 	)
 	for i := uint64(0); i < 500; i++ {
 		var txdata Transaction
@@ -577,11 +573,11 @@ func randHashes(n int) []libcommon.Hash {
 	return h
 }
 
-func randAccessList() libtypes.AccessList {
+func randAccessList() AccessList {
 	size := randIntInRange(4, 10)
-	var result libtypes.AccessList
+	var result AccessList
 	for i := 0; i < size; i++ {
-		var tup libtypes.AccessTuple
+		var tup AccessTuple
 
 		tup.Address = *randAddr()
 		tup.StorageKeys = append(tup.StorageKeys, randHash())
@@ -738,78 +734,9 @@ func TestBlobTxEncodeDecode(t *testing.T) {
 	}
 }
 
-func makeBlobTxRlp() []byte {
-	bodyRlp := hexutility.MustDecodeHex(txpool.BodyRlpHex)
-
-	blobsRlpPrefix := hexutility.MustDecodeHex("fa040008")
-	blobRlpPrefix := hexutility.MustDecodeHex("ba020000")
-
-	var blob0, blob1 = gokzg4844.Blob{}, gokzg4844.Blob{}
-	copy(blob0[:], hexutility.MustDecodeHex(txpool.ValidBlob1Hex))
-	copy(blob1[:], hexutility.MustDecodeHex(txpool.ValidBlob2Hex))
-
-	var err error
-	proofsRlpPrefix := hexutility.MustDecodeHex("f862")
-	commitment0, _ := kzg.Ctx().BlobToKZGCommitment(blob0, 0)
-	commitment1, _ := kzg.Ctx().BlobToKZGCommitment(blob1, 0)
-
-	proof0, err := kzg.Ctx().ComputeBlobKZGProof(blob0, commitment0, 0)
-	if err != nil {
-		fmt.Println("error", err)
-	}
-	proof1, err := kzg.Ctx().ComputeBlobKZGProof(blob1, commitment1, 0)
-	if err != nil {
-		fmt.Println("error", err)
-	}
-
-	wrapperRlp := hexutility.MustDecodeHex("03fa0401fe")
-	wrapperRlp = append(wrapperRlp, bodyRlp...)
-	wrapperRlp = append(wrapperRlp, blobsRlpPrefix...)
-	wrapperRlp = append(wrapperRlp, blobRlpPrefix...)
-	wrapperRlp = append(wrapperRlp, blob0[:]...)
-	wrapperRlp = append(wrapperRlp, blobRlpPrefix...)
-	wrapperRlp = append(wrapperRlp, blob1[:]...)
-	wrapperRlp = append(wrapperRlp, proofsRlpPrefix...)
-	wrapperRlp = append(wrapperRlp, 0xb0)
-	wrapperRlp = append(wrapperRlp, commitment0[:]...)
-	wrapperRlp = append(wrapperRlp, 0xb0)
-	wrapperRlp = append(wrapperRlp, commitment1[:]...)
-	wrapperRlp = append(wrapperRlp, proofsRlpPrefix...)
-	wrapperRlp = append(wrapperRlp, 0xb0)
-	wrapperRlp = append(wrapperRlp, proof0[:]...)
-	wrapperRlp = append(wrapperRlp, 0xb0)
-	wrapperRlp = append(wrapperRlp, proof1[:]...)
-
-	return wrapperRlp
-}
-
-// This test is for reference
 func TestShortUnwrap(t *testing.T) {
-	blobTxRlp := makeBlobTxRlp()
+	blobTxRlp, _ := typestest.MakeBlobTxRlp()
 	shortRlp, err := UnwrapTxPlayloadRlp(blobTxRlp)
-	if err != nil {
-		t.Errorf("short rlp stripping failed: %v", err)
-		return
-	}
-	prefixedRlp := append([]byte{0x03}, shortRlp...) // Added the 0x3 prefix for DecodeTransaction func
-	bbtx, err := DecodeTransaction(prefixedRlp)
-
-	if err != nil {
-		t.Errorf("short rlp decoding failed : %v", err)
-	}
-	wrappedBlobTx := BlobTxWrapper{}
-	err = wrappedBlobTx.DecodeRLP(rlp.NewStream(bytes.NewReader(blobTxRlp[1:]), 0))
-	if err != nil {
-		t.Errorf("long rlp decoding failed: %v", err)
-	}
-	assertEqual(bbtx, &wrappedBlobTx.Tx)
-}
-
-// This test actually applies to testing the behaviour as seen from the
-// onNewTx behaviour in filters.go
-func TestShortUnwrapLib(t *testing.T) {
-	blobTxRlp := makeBlobTxRlp()
-	shortRlp, err := libtypes.UnwrapTxPlayloadRlp(blobTxRlp)
 	if err != nil {
 		t.Errorf("short rlp stripping failed: %v", err)
 		return
@@ -820,7 +747,8 @@ func TestShortUnwrapLib(t *testing.T) {
 		t.Errorf("short rlp decoding failed : %v", err)
 	}
 	wrappedBlobTx := BlobTxWrapper{}
-	err = wrappedBlobTx.DecodeRLP(rlp.NewStream(bytes.NewReader(makeBlobTxRlp()[1:]), 0))
+	blockTxRlp2, _ := typestest.MakeBlobTxRlp()
+	err = wrappedBlobTx.DecodeRLP(rlp.NewStream(bytes.NewReader(blockTxRlp2[1:]), 0))
 	if err != nil {
 		t.Errorf("long rlp decoding failed: %v", err)
 	}
