--- core/vm/exports_test.go
@@ -1,5 +1,0 @@
-package vm
-
-func MemoryGasCost(mem *Memory, newMemSize uint64) (uint64, error) {
-	return memoryGasCost(mem, newMemSize)
-}
