+++ txnprovider/txpool/txn.go
@@ -1027,29 +1027,3 @@ func (tx *TxSlot) PrintDebug(prefix string) {
 	fmt.Printf("%s: senderID=%d,nonce=%d,tip=%d,v=%d\n", prefix, tx.SenderID, tx.Nonce, tx.Tip, tx.Value.Uint64())
 	//fmt.Printf("%s: senderID=%d,nonce=%d,tip=%d,hash=%x\n", prefix, tx.senderID, tx.nonce, tx.tip, tx.IdHash)
 }
-
-// Removes everything but the payload body from blob tx and prepends 0x3 at the beginning - no copy
-// Doesn't change non-blob tx
-func UnwrapTxPlayloadRlp(blobTxRlp []byte) ([]byte, error) {
-	if blobTxRlp[0] != BlobTxType {
-		return blobTxRlp, nil
-	}
-	dataposPrev, _, isList, err := rlp.Prefix(blobTxRlp[1:], 0)
-	if err != nil || dataposPrev < 1 {
-		return nil, err
-	}
-	if !isList { // This is clearly not wrapped txn then
-		return blobTxRlp, nil
-	}
-
-	blobTxRlp = blobTxRlp[1:]
-	// Get to the wrapper list
-	datapos, datalen, err := rlp.List(blobTxRlp, dataposPrev)
-	if err != nil {
-		return nil, err
-	}
-	blobTxRlp = blobTxRlp[dataposPrev-1 : datapos+datalen] // seekInFiles left an extra-bit
-	blobTxRlp[0] = 0x3
-	// Include the prefix part of the rlp
-	return blobTxRlp, nil
-}
